Fix spotting brackets inside in a Lambda
The previous code tried to account for `\(a -> b) -> c` - noting that `b` was not the body of a lambda, so shouldn't be marked as not requiring brackets. However, it was the body of a view pattern, so didn't need brackets, so the whole dance was pointless. Furthermore, the whole dance was wrong twice over, and the original code, with the original flawed logic, should have read:

    | Lambda{} <- parent, i == length (children parent) - 1 = False